storybook: move track occupancy diagram story and samples
Move these to storybook/stories/ui-charts/trackOccupancyDiagram/
for consistency with other stories.
