+++ storybook/stories/ui-charts/trackOccupancyDiagram/assets/occupancyZones.ts
