+++ storybook/stories/ui-charts/trackOccupancyDiagram/assets/tracks.ts
