+++ storybook/stories/ui-charts/trackOccupancyDiagram/rendering.stories.tsx
@@ -2,25 +2,25 @@ import React, { useEffect, useMemo, useState } from 'react';
 
 import type { Meta, StoryObj } from '@storybook/react';
 
-import { KebabHorizontal } from '../../../ui-icons/src/index';
-import TimeCaptions from '../../../ui-charts/src/spaceTimeChart/components/TimeCaptions';
-import { useCanvas, useDraw } from '../../../ui-charts/src/spaceTimeChart/hooks/useCanvas';
-import { useMouseInteractions } from '../../../ui-charts/src/spaceTimeChart/hooks/useMouseInteractions';
-import { useMouseTracking } from '../../../ui-charts/src/spaceTimeChart/hooks/useMouseTracking';
-import { useSize } from '../../../ui-charts/src/spaceTimeChart/hooks/useSize';
-import { DEFAULT_THEME } from '../../../ui-charts/src/spaceTimeChart/lib/consts';
+import { KebabHorizontal } from '../../../../ui-icons/src/index';
+import TimeCaptions from '../../../../ui-charts/src/spaceTimeChart/components/TimeCaptions';
+import { useCanvas, useDraw } from '../../../../ui-charts/src/spaceTimeChart/hooks/useCanvas';
+import { useMouseInteractions } from '../../../../ui-charts/src/spaceTimeChart/hooks/useMouseInteractions';
+import { useMouseTracking } from '../../../../ui-charts/src/spaceTimeChart/hooks/useMouseTracking';
+import { useSize } from '../../../../ui-charts/src/spaceTimeChart/hooks/useSize';
+import { DEFAULT_THEME } from '../../../../ui-charts/src/spaceTimeChart/lib/consts';
 import {
   CanvasContext,
   MouseContext,
   SpaceTimeChartContext,
-} from '../../../ui-charts/src/spaceTimeChart/lib/context';
+} from '../../../../ui-charts/src/spaceTimeChart/lib/context';
 import type {
   MouseContextType,
   SpaceTimeChartContextType,
   PickingElement,
   SpaceTimeChartTheme,
-} from '../../../ui-charts/src/spaceTimeChart/lib/types';
-import { OPERATIONAL_POINTS } from '../../../ui-charts/src/spaceTimeChart/stories/lib/paths';
+} from '../../../../ui-charts/src/spaceTimeChart/lib/types';
+import { OPERATIONAL_POINTS } from '../../../../ui-charts/src/spaceTimeChart/stories/lib/paths';
 import {
   getTimeToPixel,
   getSpaceToPixel,
@@ -29,13 +29,13 @@ import {
   getPixelToSpace,
   getPointToData,
   spaceScalesToBinaryTree,
-} from '../../../ui-charts/src/spaceTimeChart/utils/scales';
+} from '../../../../ui-charts/src/spaceTimeChart/utils/scales';
 import {
   TrackOccupancyManchette,
   TrackOccupancyCanvas,
-} from '../../../ui-charts/src/trackOccupancyDiagram/index';
-import occupancyZones from '../samples/TrackOccupancyDiagramSamples/occupancyZones';
-import tracks from '../samples/TrackOccupancyDiagramSamples/tracks';
+} from '../../../../ui-charts/src/trackOccupancyDiagram/index';
+import occupancyZones from './assets/occupancyZones';
+import tracks from './assets/tracks';
 
 type TrackOccupancyDiagramProps = {
   xZoomLevel: number;
